requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.2
language-tool-python>=2.7.0
textblob>=0.17.1
nltk>=3.8.1
python-dotenv>=1.0.0
google-generativeai>=0.5.0
